refactor: state locking and sync across update calls
This introduces a mutex into appstate.State and adds logic to TryLock at the start of the update function. This ensures that we're being obvious if there's multiple attempted node update calls at once.

We don't want state to drift and fall out of sync because of competing update flows. This should help address #14, but needs a bit of auditing. It also may need to be tighter scoped for locking if we're seeing issues with non-update read operations or performance impacts.