Updating everything to pass in the valid trace context as a logging field
All log calls will pass in the context object created by starting a new span. The logging core pulls the context out of the fields and retrieves the active span from that context. From there, we turn it into a ReadOnlySpan that gives us access to all of the attributes on the active span.

We add in the appropriate fields, including extra span information if debug logging is enabled, and then drop the trace context from the log fields and send it to the ioCore to log out.

This also includes a debug Dockerfile that's used to build an image that includes Delve for remote debugging.