feat: Updated observability and improved logging (#36)
* fix: Removed the the tracing span logic when reading the config.

* custom zapcore logger for injecting trace ID and span details into log messages

* chore(test): updating tests to correctly use pointers for contextvalues

* Changing how the custom logger is getting span information.

The provided context isn't going to be associated with a given span, so we need to pass in the span context to the logger as a field during each log call

taking out the span.IsValid() call

* Updating everything to pass in the valid trace context as a logging field

All log calls will pass in the context object created by starting a new span. The logging core pulls the context out of the fields and retrieves the active span from that context. From there, we turn it into a ReadOnlySpan that gives us access to all of the attributes on the active span.

We add in the appropriate fields, including extra span information if debug logging is enabled, and then drop the trace context from the log fields and send it to the ioCore to log out.

This also includes a debug Dockerfile that's used to build an image that includes Delve for remote debugging.

* fixing golangci-lint issues with ineff assignment