feat: adding support for optional drain conditions
this allows for configuring some additional drain conditions. the idea
is that the conditions would be significant enough that a node could be
rendered unstable or entirely unusable and draining applications and
workload from the nodes would be the best option for mitigating
potential impact.

because this also changes the way we check for whether or not a node
condition is part of what's considered "drainable", we also pushed
additional logic to separate a live migration freeze event from a
regular freeze event that may not have the same impact as a live
migration. this allows handling freezes differently than live
migrations.